Give NeoVim another shot
Previously, NeoVim was unusable for me because shell commands executed
via `:! ...` did not have color support and using `:terminal` replaces
the current buffer *and closes it after the command finished*! All of
that still happens, but I'm trying out workflow of mimicking the
behaviour of `:terminal` with the change of leaving me in normal mode
when the command finishes. Then the previous buffer is a `<c-^>` away.
Currently, this feels ridiculous, but maybe I'll get used to it.

In my search to figure out a good workflow for the problem above, I
stumbled on the true color support. It didn't solve my problem, but
after enabling it (and installing the iTerm nightly), my editor is
noticeably prettier. However, it doesn't Just Work with tmux, so I'm
going to abandon it for the moment. I think I'll have to deep dive into
the details of terminal color codes to really understand what the hell
is happening and why.