Don't disable spring system-wide
This is an experiment. The argument for disabling it was

> As soon as spring breaks down, you spend more time resolving whatever
> the issue is, than the amount of time spring saved you.

It's quite likely that that's true, however, I think the time difference
is negligible in the long run, but the feeling of faster tests
in-the-moment of executing then has a much bigger emotional effect and
may be worth the cost of dealing with spring's issues.