Dedicate ctrl+left (right) to word movement
The pane resize binding was tying up `control` and `shift` + arrow keys.
I only use the big resize, on `control`, so the `shift` bindings went
unused.  I'd rather use `control` for word movement on the command line,
so this change drops the small resize bindings and moves the bigger ones
to be bound on `shift`.

Also removes some bindkey nonsense that was an attempt to get `option` +
arrows working as word movement. Coincidentally, `option` now works
along with `control`. :/