Emulate `:!`
Adds a function that executes a given command in a new tab using
`:terminal`. This sorta emulates `:!` from Vim in that the command takes
over the whole screen, supports color, and is interactive. This seems
like a totally feasible alternative to `:!` and it may well be enough to
get me over the barrier that kept me from using Neovim all this time.