Introduce git ss alias as `diff --shortstat`
When working on a feature branch, I often stop to check in on my current
diff size, to judge whether I should make an attempt to submit a pull
request with the work done so far, or continue implementing the feature.
I don't have any hard limits, but usually when a diff is in the
ballpark of 300 line changes, I'll make an attempt to submit a PR with
those isolated changes.

Note: this will only calculate stats for files already in the index,
e.g. new files won't be included unless they are added with `git add -N`
first.

Note: to calculate these stats on changes that have been staged, execute
`git ss --cached`. I don't think I do this often enough to warrant an
alias for it.