Define a custom color scheme
Using `syntax off` has some downsides:

* vim-endwise stops completing Ruby blocks/methods with `end` keywords
* I lose the "ruby method" movement (is this provided by a plugin?)
* The `%` command does not support `do`/`end` in Ruby files

This change adds a custom color scheme where (most) everything is black
on white. The current exceptions that I know of are Todo comments, which
are white on black so they stand out like a sore thumb, and match
highlights, which are kept at the default yellow background. There may
be other syntax definitions I missed, but I'll tweak this as I discover
them.