Use system Ruby for `code`
The shebang in `selecta` is handled by rbenv's Ruby shim. If it's invoked
in a directory with a .ruby-version file, the shim will execute
`selecta` with that version of Ruby. If the referenced version of Ruby
is not installed on the system, it will fail.

This updates the command to always invoke `selecta` with the system
Ruby, which would always be available.