Fix `K`
Overriding the `keywordprg` variable changes the command that runs when
`K` is pressed. This works great for everything except Ruby files, where
it's overridden to use `ri`. I don't install any local gem
documentation, so this is not helpful for me. I don't feel like taking
the time to determine where this override happens so I can put a stop to
it and remapping the key does what I want.