Vim: Clear screen before executing file
When executing files from within Vim, it can be difficult to tell the
current output from the previous when they're all smashed together with
no prompt to separate them. This updates each execution mapping to clear
the terminal before execution so it's clear what output is relevant.

It would probably make sense to consolidate this in a function to
execute commands, but I don't have time to do that at the moment.