Remove ALE gutter styles
If there are linter warnings/errors in an buffer, and I trigger a spec
run or anything that uses :! before ALE finishes its async run, its
background colors will affect the output of the running command. This
often resulted in a bright yellow background, which made the output
basically impossible to read. I enjoy the async linting quite a bit, so
I want to keep it around. This removes the colors for ALE's gutter text
in an attempt to make this experience better.