ZSH: Make asdf work? I don't even know
I've made so many changes about where and when asdf is loaded and I have
no idea what it should be. It seems to have different behavior on the
different computers I use, so that's fun.

Currently, I'm having a problem where my shell sees a different version
of Ruby than Vim. These changes _seem_ to fix that problem, but I
confess it was random flailing that ended up working.

Until next time!