Git: Remove unused aliases
Once upon a time my Git workflow made heavy use of commit taking
advantage of the autosquash rebase option. The problem with it is that I
ended up not having meaningful commit messages for the intermediate
commits, which required some extra effort when squashing. I now use the
wip alias to include a message and manually squash/fixup after the fact.