Tmux: remove some bindings
I found myself not really using the `-` or `\` bindings, instead
favoring the defaults of `"` and `%`.

When Vim gets launches as a sub-process, (I don't know if this is the
correct description), e.g. `git commit`, the `C-h` and family of
bindings stopped working. I believe this is because
`pane_current_command` returns the name of the parent, which does not
contain the phrase 'vim', so the binding doesn't forward the keys to the
sub-process. Rather than attempt to figure out to fix this, I'll remove
these bindings all together and use the defaults. This means `C-h` and
etc. always work in Vim, and to navigate tmux panes I need to go through
the prefix.