tmux: enable the mouse
This is 100% so I can sometimes be lazy and scroll through output log
with the scroll wheel.

A down side is that if I want to select text with the mouse to copy with
`cmd+c`, I have to hold shift first. Otherwise tmux takes control of the
highlight and does not do the right thing.