fish: better tat function
`$TMUX` is always empty when `tat` is invoked outside of tmux. Since I
don't use `tat` from within tmux, using it as indicator of how to behave
isn't very useful. With the current logic, `tat` creates a new session
every time it's invoked.

This flips the logic to check for the session by name, attaching to it
if it exists and creating it otherwise.  If invoked within tmux, it
aborts.