Homebrew: no auto updates, self cleanup
Configures Homebrew to never update itself automatically. I have a
routine where I do this manually once a week and it's always annoying to
wait on an automatically triggered update.

Also configures Homebrew to perform a cleanup when upgrading packages.

Thanks to Teo for exposing me to Homebrew's wide array of ENV vars:
https://github.com/teoljungberg/dotfiles/commit/58036d6da3e37e407dc7c7ff3cc3e9556c8bff4e