Notes: spike porting note command to Ruby
The previous versions of this command opened a temp file in the editor.
The idea for that was that I get a "clean" file each time, so new notes
aren't affected by the contents of previous ones (for the same tag and
day). I don't think that was ever that valuable to me, so this version
opens the note file directly, allowing for creating and editing with the
same command (again only for the current day).

For now, that's all this version supports.

Shebang note: I'm specifically targeting the system ruby (at least on
Mac) because using asdf's ruby has a noticeable launch time. I may look
into this further at some point, but for now I just want it to be fast.

```
% time /usr/bin/ruby -v
ruby 2.3.7p456 (2018-03-28 revision 63024) [universal.x86_64-darwin18]
        0.02 real         0.01 user         0.00 sys

% time /usr/bin/env ruby -v
ruby 2.5.0p0 (2017-12-25 revision 61468) [x86_64-darwin16]
        0.27 real         0.10 user         0.14 sys

% time ~/.asdf/shims/ruby -v
ruby 2.5.0p0 (2017-12-25 revision 61468) [x86_64-darwin16]
        0.25 real         0.10 user         0.13 sys
```