Alacritty: initial config
I'm going to give Alacritty a trial run. The primary appeal here is that
it aims to be simple and let multiplexers handle tabs/splits/etc.
Kitty's anti-multiplexer stance has caused me some grief and I'm curious
to see how Alacritty does.

This adds the full default config with comments removed and some tweaks
I made while reading through it. Deviations from the defaults in a
nutshell:

* font and colors copied from Kitty config
* hide mouse while typing
* cmd+click to open urls
* buttonless window so there's no title bar
* added mappings for alt+b and alt+f to move back and forward by word

It's a bummer that all the key bindings have to be specified here. I
mean, it's cool that you can see the full list of all the bindings, but
having 100+ lines in my config is annoying.