tmux: tweak ctr-p
`choose-tree` shows not only the sessions to switch to, but also each
window and split within the windows for every session. Coupled with my
`tat` function, having 2 tmux sessions puts 8 entries in the
`choose-tree` menu, 1 for each overall session and 1 for each of the 3
windows in each session. It's a lot of noise and makes using the key
shortcuts for them (e.g. 0-9, etc) difficult because of the noise.

This switches to `choose-session`, which only includes 1 entry for each
session.

I also learned that by default `choose-tree` is bound to `<prefix>-w`,
so I can use that if I do need it in the future.