Alacritty: "Forward" all alt+alphabet keys
Similar to c7a5479, I've found a handful of alt combinations that didn't
do the right thing. The main examples being alt+{h,j,k,l} for switching
tmux panes. In a I-am-tired-of-dealing-with-this-fashion, this commit
adds mappings for every alphabet character with alt. I'm no scientist,
but I'm thinking of these as "forwards," which let's the shell decide
what to do, instead of letting Alacritty handle it. It seems to be
working ¯\(°_o)/¯