Fish: why are paths so hard?
Fish 3.0 now treats the PATH much like masOS's `path_helper` utility (or
it uses it directly, I'm not sure). This totally broke all my
environments, because it gives `/etc/paths` priority, making the system
`ruby` and `python` executables more visible than asdf's (or any package
manager, e.g. homebrew, rbenc, etc).

To get around this, it's recommended to modify the `$fish_user_paths`
universal variable instead of manipulating `$PATH` directly. The
contents of this variable are prepended to `$PATH` when a new shell is
launched, letting us add higher priority entries than `/etc/paths`.

Since `$fish_user_paths` is a universal variable, setting it once will
apply it... universally, for now and all future shells. That makes it
difficult to manage via the config script, because the value is
persisted. Adding an entry to it in the config will prepend it to the
existing value, resulting in duplicate entries. If you launch another
shell, there'll be three duplicate entries, etc.

Thus, this change first checks if the entry I want to add to the path
exists and will only add it if it does not. This appears to be working,
though there is an edge case somewhere that still results in duplicate
entries when I'm using `pipenv shell` that I have not been able to track
down. I'm beyond frustrated with it so I'm going to abandon using
`pipenv shell` entirely and use `pipenv run *` for the things that need
it.