Fish: simplify prompt
Removes the current time and git branch and collapses the whole thing to
a single line.

The idea with adding the current time was that when a process took
awhile, I could get an idea of just how long without needing to run it
again with `time (1)`. In practice, this did not happen often enough to
warrant the screen real estate cost.

I do look at the git branch fairly often, but I also manually check the
git status very often. I don't think I'll miss that information.