Fish: add the current git branch to the prompt
Since removing the branch name in f17eb77, two days ago, I've
accidentally pushed to master twice. Perhaps having it always visible
was more critical than I realized. Whoops.

The `white` color for me is closer to gray, so this ends up being actual
white on top of gray. It wouldn't pass an accessibility test, but this
is just for me and I think I want the info a bit subdued.