Revert "Kitty: Hide the mouse after 2 seconds"
Turns out this is still quite annoying. Most of the time it's fine, but
sometime Kitty won't show the mouse after it loses focus, so I'm left
without being able to see it when I need to. Restarting Kitty seems to
fix it for a bit, but it's frustrating. I'd rather deal with always
having the mouse.

This reverts commit 956868e93d32f50940d5a47c06dc8106f8ccc186.