Vim: Remove language plugins that vanilla Vim has
The scripts that ship with Vim are still loaded for each file they're
relevant to, meaning there's some duplicated work happening. I don't
think I've noticed any slowness in loading because of this, but knowing
it's happening does bother me a bit.

I think the versions in Vim proper lag behind a bit, but I'm not sure
that matters. If I encounter that situation, I'll re-add individual
plugins.