Vim: Switch from ctags.vim to Gutentags
Imma be honest here, I'm not totally sure why I have ctags.vim or what
it's doing for me. Seemingly, it's doing nothing. Perhaps this was
because I'm missing some config or some other user error. Either way,
I've had this plugin around for years and I can't say I noticed it.

Regardless, I saw mention of Gutentags (shout out to @mike-burns) and I
gave it a shot and was pleased with the auto tag generation, so I'm
switching to it.