Vim: start tracking spell additions
Vim uses the binary .spl file to spell check at runtime. It tracks your
additions in a text file, and builds the .spl file with `mkspell` as
things are added. Since I'm putting my add file in Git, it's possible
that file will change without Vim's knowledge. To support that, we check
if the modified time of both files and rebuild the .spl if needed.