Vim: better semantic execution for ruby scripts
Defining the mapping at buffer load means opening a new file will always
have the ruby mapping, even if the rails runner comment is added later.
Currently have to reload the buffer with `:e` for the mapping to change.

With a function, we'll make this check every time the mapping is used,
so we can swap between the execution types without reloading.