Revert "Fish: use built in path helper"
This seems to have caused the .git/safe/../../bin entry to globally and
permanently add the expanded directories to my path. E.g. `cd` into a
directory with a safe repo, start tmux, and the absolute path to that
repo's bin directory is permanently at the front of `$PATH`.

I was able to remove the offenders one at a time by alternating between
`echo $PATH` and `set -e fish_user_paths[1]`.

This was an infuriating bug to track down.

This reverts commit 0d0a9cf6cb1b87023a8b08574505d9cc83fe9ead.