Vim: Improve SQL execution
Currently, `<leader>x` will execute the current SQL buffer against the
local database, assuming a .dbname file exists and we're using Postgres.
To get around the Postgres assumption and to support the idea of local
vs remote, we now look for .dblocal and .dbremote files, which should be
executable scripts that know how to connect to whichever database is
appropriate. The scripts receive the path to the SQL file as an arg.

`<leader>x` is local, `<leader>X` is remote.