Vim: force + register to system clipboard
On macos I haven't had to consider the clipboard option before, the +
register always worked how I wanted. On linux, the default value is
allowing me to paste from the system clipboard with `"+p`, but yanking
with `"+y` does not copy to the system clipboard. `unnamedplus` seems to
resolve that.