Git: update `sha` to also print the sha
Accomplishing this is pretty awkward in Fish. In other shells, this
could be done with process substitution as input (I'm unsure of this
terminology), e.g. in Bash `echo 'foo' | tee >(pbcopy)`. Fish does not
have process substitution in this direction, so we workaround by
utilizing the tty file descriptor.

A discussion about adding this functionality to Fish:

    https://github.com/fish-shell/fish-shell/issues/1786