Fish: fix fish_features variable
Not sure when this happened, but these features are no longer being set
for sessions. I wanted to avoid the `-U` universal option, since the
whole point of the config file is to set things up on demand, whereas
universal intended to be a set once, last forever kinda thing. I tried
using the `-x` export option like the other variables here, but that
didn't seem to work. Checked with the `set -S fish_features` command.