Git: remove m and supporting aliases
I haven't used this in quite a while, favoring the squash and merge
button on GitHub. Also, the `main-local` alias doesn't seem to work
anymore. I briefly tried to fix it, but because there is no concept of a
default branch in git itself, it seemed the only way to get this value
is from the origin info, which requires talking to origin and is too
slow for my tastes. For posterity, something like this would get the
correct value:

```
git remote show origin | grep 'HEAD branch' | cut -w -f4
```