Git: bring back mail-local
This was removed in de61610, citing there was no way to detect the main
branch, because that's a GitHub idea, not a Git idea. I actually do want
this functionality though, for other scripting reasons, so it comes back
here without autodetection. It will look for a _.git/main-local_ file,
which it assumes will have the name of the main branch and only that,
and fall back to `main` if no file exists.